requirements.txt
aiocache==0.2.2
aiofiles==0.3.0
aiomcache==0.4.0
aioredis==0.2.9
aiosmtpd==1.0a4
atpublic==0.5
boto3==1.4.4
botocore==1.5.2
cassandra-driver==3.7.1
docutils==0.13.1
hiredis==0.2.0
httptools==0.0.9
jmespath==0.9.0
multidict==2.1.4
python-dateutil==2.6.0
redis==2.10.5
requests==2.12.4
s3transfer==0.1.10
Sanic==0.1.8
six==1.10.0
ujson==1.35
uvloop==0.6.7
validate-email==1.3
